+++ src/app/advertiser/advertiser.module.ts
@@ -6,13 +6,15 @@ import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatDatepickerModule } from '@angular/material/datepicker';
 import { MatMomentDateModule } from '@angular/material-moment-adapter';
-import { MatInputModule } from '@angular/material';
+import { MatCheckboxModule, MatInputModule } from '@angular/material';
 import { MatSelectModule } from '@angular/material/select';
 import { MatExpansionModule } from '@angular/material/expansion';
 import { MatMenuModule } from '@angular/material/menu';
+import { MatTabsModule } from '@angular/material/tabs';
 import { FileUploadModule } from 'ng2-file-upload';
-import { OwlDateTimeModule, OwlNativeDateTimeModule, OWL_DATE_TIME_FORMATS } from 'ng-pick-datetime';
+import { OWL_DATE_TIME_FORMATS, OwlDateTimeModule, OwlNativeDateTimeModule } from 'ng-pick-datetime';
 import { OwlMomentDateTimeModule } from 'ng-pick-datetime-moment';
+import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 import { AppCommonModule } from 'common/common.module';
 import { AdvertiserComponent } from './advertiser.component';
 import { AdvertiserRoutingModule } from './advertiser-routing.module';
@@ -20,6 +22,7 @@ import { EditCampaignComponent } from './edit-campaign/edit-campaign.component';
 import { CampaignListComponent } from './campaign-list/campaign-list.component';
 import { CampaignListItemComponent } from './campaign-list/campaign-list-item/campaign-list-item.component';
 import { EditCampaignBasicInformationComponent } from './edit-campaign/edit-campaign-basic-info/edit-campaign-basic-information.component';
+import { EditCampaignConversionComponent } from './edit-campaign/edit-campaign-conversion/edit-campaign-conversion.component';
 import { EditCampaignAdditionalTargetingComponent } from './edit-campaign/edit-campaign-additional-targeting/edit-campaign-additional-targeting.component';
 import { EditCampaignCreateAdsComponent } from './edit-campaign/edit-campaign-create-ads/edit-campaign-create-ads.component';
 import { EditCampaignSummaryComponent } from './edit-campaign/edit-campaign-summary/edit-campaign-summary.component';
@@ -28,32 +31,32 @@ import { CampaignDetailsComponent } from './campaign-details/campaign-details.co
 import { AdListComponent } from './campaign-details/ad-list/ad-list.component';
 import { AdListItemComponent } from './campaign-details/ad-list/ad-list-item/ad-list-item.component';
 
-import { CampaignBudgetPerDayPipe } from "common/pipes/campaign-budget-per-day.pipe";
+import { CampaignBudgetPerDayPipe } from 'common/pipes/campaign-budget-per-day.pipe';
 import { AdvertiserGuard } from './advertiser-guard.service';
 import { CampaignResolver } from './resolvers/campaign.resolver';
 import { TargetingCriteriaResolver } from './resolvers/targeting-criteria.resolver';
-import {
-  DATE_AND_TIME_PICKER_FORMATS
-} from "common/utilities/consts";
-
+import { DATE_AND_TIME_PICKER_FORMATS } from 'common/utilities/consts';
 
 
 const matModules = [
+  MatCheckboxModule,
   MatExpansionModule,
   MatFormFieldModule,
   MatMomentDateModule,
   MatDatepickerModule,
   MatInputModule,
   MatMenuModule,
-  MatSelectModule
+  MatSelectModule,
+  MatTabsModule,
 ];
 
 const editCampaignComponents = [
   EditCampaignComponent,
   EditCampaignBasicInformationComponent,
+  EditCampaignConversionComponent,
   EditCampaignAdditionalTargetingComponent,
   EditCampaignCreateAdsComponent,
-  EditCampaignSummaryComponent
+  EditCampaignSummaryComponent,
 ];
 
 const advertiserComponents = [
@@ -63,7 +66,7 @@ const advertiserComponents = [
   DashboardComponent,
   CampaignDetailsComponent,
   AdListComponent,
-  AdListItemComponent
+  AdListItemComponent,
 ];
 
 @NgModule({
@@ -72,25 +75,26 @@ const advertiserComponents = [
     HttpModule,
     AppCommonModule,
     AdvertiserRoutingModule,
+    FontAwesomeModule,
     FormsModule,
     ReactiveFormsModule,
     FileUploadModule,
     OwlDateTimeModule,
     OwlNativeDateTimeModule,
     OwlMomentDateTimeModule,
-    ...matModules
+    ...matModules,
   ],
   providers: [
     AdvertiserGuard,
     CampaignResolver,
     TargetingCriteriaResolver,
-    {provide: OWL_DATE_TIME_FORMATS, useValue: DATE_AND_TIME_PICKER_FORMATS}
+    {provide: OWL_DATE_TIME_FORMATS, useValue: DATE_AND_TIME_PICKER_FORMATS},
   ],
 
   declarations: [
     CampaignBudgetPerDayPipe,
     ...advertiserComponents,
-    ...editCampaignComponents
+    ...editCampaignComponents,
   ]
 })
 export class AdvertiserModule {