+++ src/app/models/initial-state/ad-unit.ts
@@ -1,9 +1,9 @@
 import { AdUnit } from 'models/site.model';
-import { adTypesOptions, adUnitStatusesEnum } from 'models/enum/ad.enum';
+import { adUnitTypesEnum, adUnitStatusesEnum } from 'models/enum/ad.enum';
 
 
 export const adUnitInitialState: Partial<AdUnit> = {
-  shortHeadline: '',
-  type: adTypesOptions[0],
+  name: '',
+  type: adUnitTypesEnum.DISPLAY,
   status: adUnitStatusesEnum.ACTIVE,
 };