+++ src/app/publisher/edit-site/edit-site-basic-info/edit-site-basic-information.component.ts
@@ -23,6 +23,7 @@ import { faQuestionCircle } from '@fortawesome/free-solid-svg-icons';
 })
 export class EditSiteBasicInformationComponent extends HandleSubscription implements OnInit {
   private static readonly WEBSITE_NAME_LENGTH_MAX: number = 64;
+  private static readonly WEBSITE_DOMAIN_LENGTH_MAX: number = 255;
   faQuestionCircle = faQuestionCircle;
   siteBasicInfoForm: FormGroup;
   languages: SiteLanguage[];
@@ -31,6 +32,8 @@ export class EditSiteBasicInformationComponent extends HandleSubscription implem
   createSiteMode: boolean;
   filteredOptions: Observable<object>;
   changesSaved: boolean = false;
+  websiteNameLengthMax = EditSiteBasicInformationComponent.WEBSITE_NAME_LENGTH_MAX;
+  websiteDomainLengthMax = EditSiteBasicInformationComponent.WEBSITE_DOMAIN_LENGTH_MAX;
 
   constructor(
     private router: Router,
@@ -83,7 +86,7 @@ export class EditSiteBasicInformationComponent extends HandleSubscription implem
     this.store.dispatch(new SaveLastEditedSite(this.site));
     this.changesSaved = true;
     this.router.navigate(
-      ['/publisher', 'create-site', 'additional-filtering'],
+        ['/publisher', 'create-site', 'additional-filtering'],
       {queryParams: {step: 2}}
     );
   }
@@ -108,8 +111,21 @@ export class EditSiteBasicInformationComponent extends HandleSubscription implem
     this.site = {
       ...this.site,
       name: this.siteBasicInfoForm.controls['name'].value,
+      domain: this.siteBasicInfoForm.controls['domain'].value,
       primaryLanguage: typeof chosenLanguage === 'object' ? chosenLanguage.code : chosenLanguage,
     };
+  }
+
+  extractDomain(): void {
+    const url = this.siteBasicInfoForm.get('domain').value;
+
+    let domain = url.toLowerCase();
+    //remove protocol, user info and www subdomain
+    domain = domain.replace(/^(?:[a-z0-9+.-]+:\/\/)?(?:\/\/)?(?:.*@)?(?:www\.)?/i, "");
+    // remove port number, path, query string and fragment
+    domain = domain.replace(/(?::.*)?(?:\/.*)?(?:\?.*)?(?:#.*)?$/i, "");
+
+    this.siteBasicInfoForm.get('domain').setValue(domain);
   }
 
   updateSite(): void {
@@ -126,6 +142,11 @@ export class EditSiteBasicInformationComponent extends HandleSubscription implem
       name: new FormControl(siteInitialState.name, [
         Validators.required,
         Validators.maxLength(EditSiteBasicInformationComponent.WEBSITE_NAME_LENGTH_MAX)
+      ]),
+      domain: new FormControl(siteInitialState.domain, [
+        Validators.required,
+        Validators.maxLength(EditSiteBasicInformationComponent.WEBSITE_DOMAIN_LENGTH_MAX),
+        Validators.pattern('^.+\..+$')
       ]),
       primaryLanguage: new FormControl(siteInitialState.primaryLanguage, Validators.required)
     });