+++ src/app/publisher/resolvers/ad-unit-sizes.resolver.ts
@@ -3,14 +3,14 @@ import { ActivatedRouteSnapshot, Resolve } from '@angular/router';
 import { Observable } from 'rxjs/Observable';
 
 import { PublisherService } from 'publisher/publisher.service';
-import { AdUnitSize } from 'models/site.model';
+import { AdUnitMetaData } from 'models/site.model';
 
 @Injectable()
-export class AdUnitSizesResolver implements Resolve<AdUnitSize[]> {
+export class AdUnitSizesResolver implements Resolve<AdUnitMetaData[]> {
   constructor(private publisherService: PublisherService) {
   }
 
-  resolve(route: ActivatedRouteSnapshot): Observable<AdUnitSize[]> {
+  resolve(route: ActivatedRouteSnapshot): Observable<AdUnitMetaData[]> {
     return this.publisherService.getAdUnitSizes();
   }
 }