+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.0] - 2019-11-22
+### Added
+- Conversions definition
+- Target URL placeholders
+- Pops ad units configuration
+- Site domain requirement
+### Changed
+- Hid Requires section from site filtering
+
 ## [1.2.5] - 2019-10-01
 ### Changed
 - Upgrade @angular/cli to version 1.7.0
@@ -133,7 +142,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Advertiser features (Campaigns & Ads)
 - Publisher features (Sites & AdUnits)
 
-[Unreleased]: https://github.com/adshares/adpanel/compare/v1.2.5...develop
+[Unreleased]: https://github.com/adshares/adpanel/compare/v1.4.0...develop
+[1.4.0]: https://github.com/adshares/adpanel/compare/v1.2.5...v1.4.0
 [1.2.5]: https://github.com/adshares/adpanel/compare/v1.2.4...v1.2.5
 [1.2.4]: https://github.com/adshares/adpanel/compare/v1.2.1...v1.2.4
 [1.2.1]: https://github.com/adshares/adpanel/compare/v0.11.0...v1.2.1